% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/M4comp.R
\docType{package}
\name{M4comp}
\alias{M4comp}
\alias{M4comp-package}
\title{The M4comp  package provides three categories of important functions:
foo, bar and baz.}
\description{
The M4comp  package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

