% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M4comp.R
\docType{package}
\name{M4comp}
\alias{M4comp}
\alias{M4comp-package}
\title{Data from the M4 time series forecasting competition}
\source{
\url{http://forecasters.org/resources/time-series-data/}
}
\description{
The M4comp  package provides the 10.000 time series from the M4 time series forecasting competition.
}
\section{Foo functions}{

The foo functions ...
}

